+++ http4k-contract/src/main/kotlin/org/http4k/contract/openapi/v3/jacksonExt.kt
@@ -28,7 +28,8 @@ fun OpenApi3(
 fun AutoJsonToJsonSchema(json: ConfigurableJackson) = AutoJsonToJsonSchema(
     json,
     FieldRetrieval.compose(
-        SimpleLookup(metadataRetrievalStrategy = JacksonFieldMetadataRetrievalStrategy),
+        SimpleLookup(metadataRetrievalStrategy =
+        JacksonFieldMetadataRetrievalStrategy.then(PrimitivesFieldMetadataRetrievalStrategy)),
         FieldRetrieval.compose(JacksonJsonPropertyAnnotated, JacksonJsonNamingAnnotated(json))
     )
 )
@@ -74,7 +75,8 @@ class JacksonJsonNamingAnnotated(private val json: ConfigurableJackson = Jackson
 
 object JacksonFieldMetadataRetrievalStrategy : FieldMetadataRetrievalStrategy {
     override fun invoke(target: Any, fieldName: String): FieldMetadata =
-        FieldMetadata(description = target.javaClass.findPropertyDescription(fieldName))
+        FieldMetadata(target.javaClass.findPropertyDescription(fieldName)?.let { mapOf("description" to it) }
+            ?: emptyMap())
 
     private fun Class<Any>.findPropertyDescription(name: String): String? =
         kotlin.constructors.firstOrNull()?.let {