+++ http4k-security/oauth/src/main/kotlin/org/http4k/security/OAuthCallback.kt
@@ -1,9 +1,17 @@
 package org.http4k.security
 
+import dev.forkhandles.result4k.Failure
+import dev.forkhandles.result4k.Success
+import dev.forkhandles.result4k.flatMap
+import dev.forkhandles.result4k.get
+import dev.forkhandles.result4k.map
+import dev.forkhandles.result4k.mapFailure
 import org.http4k.core.HttpHandler
 import org.http4k.core.Request
 import org.http4k.core.Response
+import org.http4k.core.Status
 import org.http4k.core.Status.Companion.TEMPORARY_REDIRECT
+import org.http4k.security.oauth.server.AuthorizationCode
 import org.http4k.security.openid.IdToken
 import org.http4k.security.openid.IdTokenConsumer
 
@@ -13,34 +21,77 @@ class OAuthCallback(
     private val accessTokenFetcher: AccessTokenFetcher
 ) : HttpHandler {
 
-    override fun invoke(request: Request) = request.queryOrFragmentParameter("code")
-        ?.let { code ->
-            val state = request.queryOrFragmentParameter("state")
-            state?.let(::CrossSiteRequestForgeryToken)
-                ?.takeIf { it == oAuthPersistence.retrieveCsrf(request) }
-                ?.let {
-                    val idToken = request.queryOrFragmentParameter("id_token")?.let { IdToken(it) }
-                    if (hasValidNonceInIdToken(request, idToken)) {
-                        idToken?.let { idTokenConsumer.consumeFromAuthorizationResponse(it) }
-                        accessTokenFetcher.fetch(code)
-                            ?.let { tokenDetails ->
-                                tokenDetails.idToken?.also(idTokenConsumer::consumeFromAccessTokenResponse)
-                                val originalUri = oAuthPersistence.retrieveOriginalUri(request)?.toString() ?: "/"
-                                oAuthPersistence.assignToken(request, Response(TEMPORARY_REDIRECT)
-                                    .header("Location", originalUri), tokenDetails.accessToken, tokenDetails.idToken)
-                            }
-                    } else {
-                        null
-                    }
-                }
-        }
-        ?: oAuthPersistence.authFailureResponse()
+    override fun invoke(request: Request) = request.callbackParameters()
+        .flatMap { parameters -> validateCsrf(parameters, request, oAuthPersistence.retrieveCsrf(request)) }
+        .flatMap { parameters -> validateNonce(parameters, oAuthPersistence.retrieveNonce(request)) }
+        .flatMap { parameters -> consumeIdToken(parameters) }
+        .flatMap { parameters -> fetchAccessToken(parameters) }
+        .flatMap { tokenDetails -> consumeIdToken(tokenDetails) }
+        .map { tokenDetails ->
+            oAuthPersistence.assignToken(
+                request,
+                redirectionResponse(request),
+                tokenDetails.accessToken,
+                tokenDetails.idToken
+            )
+        }.mapFailure { oAuthPersistence.authFailureResponse(it) }.get()
 
-    private fun hasValidNonceInIdToken(request: Request, idToken: IdToken?): Boolean {
-        return if (idToken != null) {
-            idTokenConsumer.nonceFromIdToken(idToken) == oAuthPersistence.retrieveNonce(request)
-        } else true
+    private fun consumeIdToken(tokenDetails: AccessTokenDetails) =
+        tokenDetails.idToken?.let(idTokenConsumer::consumeFromAccessTokenResponse)?.map { tokenDetails }
+            ?: Success(tokenDetails)
+
+    private fun consumeIdToken(parameters: CallbackParameters) =
+        parameters.idToken?.let(idTokenConsumer::consumeFromAuthorizationResponse)?.map { parameters }
+            ?: Success(parameters)
+
+    private fun Request.callbackParameters() = authorizationCode().map {
+        CallbackParameters(
+            code = it,
+            state = queryOrFragmentParameter("state")?.let(::CrossSiteRequestForgeryToken),
+            idToken = queryOrFragmentParameter("id_token")?.let(::IdToken)
+        )
     }
+
+    private fun Request.authorizationCode() = queryOrFragmentParameter("code")?.let(::AuthorizationCode)
+        ?.let(::Success) ?: Failure(OAuthCallbackError.AuthorizationCodeMissing)
+
+    private fun validateCsrf(
+        parameters: CallbackParameters,
+        request: Request,
+        persistedToken: CrossSiteRequestForgeryToken?
+    ) = request.queryOrFragmentParameter("state")?.let(::CrossSiteRequestForgeryToken)
+        .let {
+            if (it == persistedToken) Success(parameters)
+            else Failure(OAuthCallbackError.InvalidCsrfToken(persistedToken?.value, it?.value))
+        }
+
+    private fun validateNonce(parameters: CallbackParameters, storedNonce: Nonce?) =
+        parameters.idToken?.let { idToken ->
+            val received = idTokenConsumer.nonceFromIdToken(idToken)
+            if (received == storedNonce)
+                Success(parameters) else Failure(OAuthCallbackError.InvalidNonce(storedNonce?.value, received?.value))
+        } ?: Success(parameters)
+
+    private fun fetchAccessToken(parameters: CallbackParameters) =
+        accessTokenFetcher.fetch(parameters.code.value)
+
+    private fun redirectionResponse(request: Request) = Response(TEMPORARY_REDIRECT)
+        .header("Location", oAuthPersistence.retrieveOriginalUri(request)?.toString() ?: "/")
+
+    private fun Request.queryOrFragmentParameter(name: String) = query(name) ?: fragmentParameter(name)
+
+    private data class CallbackParameters(
+        val code: AuthorizationCode,
+        val state: CrossSiteRequestForgeryToken?,
+        val idToken: IdToken?
+    )
 }
 
-private fun Request.queryOrFragmentParameter(name: String): String? = this.query(name) ?: fragmentParameter(name)
+sealed class OAuthCallbackError {
+    object AuthorizationCodeMissing : OAuthCallbackError()
+    data class InvalidCsrfToken(val expected: String?, val received: String?) : OAuthCallbackError()
+    data class InvalidNonce(val expected: String?, val received: String?) : OAuthCallbackError()
+    data class InvalidAccessToken(val reason: String) : OAuthCallbackError()
+    data class InvalidIdToken(val reason: String) : OAuthCallbackError()
+    data class CouldNotFetchAccessToken(val status: Status, val reason: String) : OAuthCallbackError()
+}