+++ http4k-security/oauth/src/test/kotlin/org/http4k/security/OAuthProviderTest.kt
@@ -3,6 +3,9 @@ package org.http4k.security
 import com.natpryce.hamkrest.and
 import com.natpryce.hamkrest.assertion.assertThat
 import com.natpryce.hamkrest.equalTo
+import dev.forkhandles.result4k.Failure
+import dev.forkhandles.result4k.Result
+import dev.forkhandles.result4k.Success
 import org.http4k.core.Credentials
 import org.http4k.core.Method.GET
 import org.http4k.core.Request
@@ -19,31 +22,47 @@ import org.http4k.core.then
 import org.http4k.hamkrest.hasBody
 import org.http4k.hamkrest.hasHeader
 import org.http4k.hamkrest.hasStatus
+import org.http4k.hamkrest.hasStatusDescription
+import org.http4k.security.ResponseType.CodeIdToken
+import org.http4k.security.openid.IdToken
+import org.http4k.security.openid.IdTokenConsumer
 import org.http4k.security.openid.RequestJwtContainer
 import org.http4k.security.openid.RequestJwts
 import org.junit.jupiter.api.Test
 
 class OAuthProviderTest {
     private val providerConfig = OAuthProviderConfig(
-        Uri.of("http://authHost"),
+        Uri.of("http://authHost/base"),
         "/auth",
         "/token",
         Credentials("user", "password"),
-        Uri.of("http://apiHost")
+        Uri.of("http://apiHost/api/")
     )
 
     private val oAuthPersistence = FakeOAuthPersistence()
 
-    private fun oAuth(persistence: OAuthPersistence, status: Status = OK, responseType: ResponseType = ResponseType.Code): OAuthProvider = OAuthProvider(
+    private fun oAuth(
+        persistence: OAuthPersistence,
+        status: Status = OK,
+        responseType: ResponseType = ResponseType.Code,
+        nonceFromIdToken:Nonce? = null,
+        resultIdTokenFromAuth:  Result<Unit, OAuthCallbackError.InvalidIdToken> = Success(Unit),
+        resultIdTokenFromAccessToken:  Result<Unit, OAuthCallbackError.InvalidIdToken> = Success(Unit)
+    ): OAuthProvider = OAuthProvider(
         providerConfig,
-        { Response(status).body("access token goes here") },
+        { Response(status).body("access token goes here").header("request-uri", it.uri.toString()) },
         Uri.of("http://callbackHost/callback"),
         listOf("scope1", "scope2"),
         persistence,
         { it.query("response_mode", "form_post") },
         { CrossSiteRequestForgeryToken("randomCsrf") },
         { Nonce("randomNonce") },
-        responseType
+        responseType,
+        idTokenConsumer = object:IdTokenConsumer{
+            override fun nonceFromIdToken(idToken: IdToken) = nonceFromIdToken
+            override fun consumeFromAuthorizationResponse(idToken: IdToken) = resultIdTokenFromAuth
+            override fun consumeFromAccessTokenResponse(idToken: IdToken) = resultIdTokenFromAccessToken
+        }
     )
 
     @Test
@@ -54,13 +73,13 @@ class OAuthProviderTest {
 
     @Test
     fun `filter - when no accessToken value present, request is redirected to expected location`() {
-        val expectedHeader = """http://authHost/auth?client_id=user&response_type=code&scope=scope1+scope2&redirect_uri=http%3A%2F%2FcallbackHost%2Fcallback&state=randomCsrf&response_mode=form_post"""
+        val expectedHeader = """http://authHost/base/auth?client_id=user&response_type=code&scope=scope1+scope2&redirect_uri=http%3A%2F%2FcallbackHost%2Fcallback&state=randomCsrf&response_mode=form_post"""
         assertThat(oAuth(oAuthPersistence).authFilter.then { Response(OK) }(Request(GET, "/")), hasStatus(TEMPORARY_REDIRECT).and(hasHeader("Location", expectedHeader)))
     }
 
     @Test
     fun `filter - accepts custom request JWT container`() {
-        val expectedHeader = """http://authHost/auth?client_id=user&response_type=code&scope=scope1+scope2&redirect_uri=http%3A%2F%2FcallbackHost%2Fcallback&state=randomCsrf&request=myCustomJwt&response_mode=form_post"""
+        val expectedHeader = """http://authHost/base/auth?client_id=user&response_type=code&scope=scope1+scope2&redirect_uri=http%3A%2F%2FcallbackHost%2Fcallback&state=randomCsrf&request=myCustomJwt&response_mode=form_post"""
 
         val jwts = RequestJwts { _, _, _ -> RequestJwtContainer("myCustomJwt") }
         assertThat(oAuth(oAuthPersistence).authFilter(jwts).then { Response(OK) }(Request(GET, "/")), hasStatus(TEMPORARY_REDIRECT).and(hasHeader("Location", expectedHeader)))
@@ -68,7 +87,7 @@ class OAuthProviderTest {
 
     @Test
     fun `filter - request redirecttion may use other response_type`() {
-        assertThat(oAuth(oAuthPersistence, OK, ResponseType.CodeIdToken)
+        assertThat(oAuth(oAuthPersistence, OK, CodeIdToken)
             .authFilter.then { Response(OK) }(Request(GET, "/")), hasStatus(TEMPORARY_REDIRECT).and(hasHeader("Location", ".*response_type=code\\+id_token.*".toRegex())))
     }
 
@@ -80,20 +99,17 @@ class OAuthProviderTest {
 
     @Test
     fun `callback - when invalid inputs passed, we get forbidden with cookie invalidation`() {
-        val invalidation = Response(FORBIDDEN)
+        assertThat(oAuth(oAuthPersistence).callback(base), hasStatus(FORBIDDEN) and hasStatusDescription("Authorization code missing"))
 
-        assertThat(oAuth(oAuthPersistence).callback(base), equalTo(invalidation))
+        assertThat(oAuth(oAuthPersistence).callback(withCookie), hasStatus(FORBIDDEN) and hasStatusDescription("Authorization code missing"))
 
-        assertThat(oAuth(oAuthPersistence).callback(withCookie), equalTo(invalidation))
-
-        assertThat(oAuth(oAuthPersistence).callback(withCode), equalTo(invalidation))
-
-        assertThat(oAuth(oAuthPersistence).callback(withCodeAndInvalidState), equalTo(invalidation))
+        assertThat(oAuth(oAuthPersistence).callback(withCodeAndInvalidState), hasStatus(FORBIDDEN) and hasStatusDescription("Invalid state (expected: null, received: notreal)"))
     }
 
     @Test
     fun `when api returns bad status`() {
-        assertThat(oAuth(oAuthPersistence, INTERNAL_SERVER_ERROR).callback(withCodeAndValidState), equalTo(Response(FORBIDDEN)))
+        oAuthPersistence.assignCsrf(Response(OK), CrossSiteRequestForgeryToken("randomCsrf"))
+        assertThat(oAuth(oAuthPersistence, INTERNAL_SERVER_ERROR).callback(withCodeAndValidState), hasStatus(FORBIDDEN) and hasStatusDescription("Failed to fetch access token (status: 500 Internal Server Error)"))
     }
 
     @Test
@@ -119,5 +135,25 @@ class OAuthProviderTest {
             .header("action", "assignToken")
 
         assertThat(oAuth(oAuthPersistence).callback(withCodeAndValidState), equalTo(validRedirectToRoot))
+    }
+
+    @Test
+    fun `api - uses base api uri`(){
+        val response = oAuth(oAuthPersistence).api(Request(GET, "/some-resource"))
+
+        assertThat(response, hasHeader("request-uri", equalTo("http://apiHost/api/some-resource")))
+    }
+
+    @Test
+    fun `id token - can fail from id_token from callback request`(){
+        val oauth = oAuth(oAuthPersistence, responseType = CodeIdToken, resultIdTokenFromAuth = Failure(OAuthCallbackError.InvalidIdToken("some reason")))
+
+        assertThat(oauth.callback(withCodeAndValidState), hasStatus(FORBIDDEN))
+    }
+
+    @Test
+    fun `id token - can fail from id_token from access token response`(){
+        val oauth = oAuth(oAuthPersistence, responseType = CodeIdToken, resultIdTokenFromAccessToken = Failure(OAuthCallbackError.InvalidIdToken("some reason")))
+        assertThat(oauth.callback(withCodeAndValidState), hasStatus(FORBIDDEN))
     }
 }