+++ http4k-testing/webdriver/src/main/kotlin/org/http4k/webdriver/Http4kWebDriver.kt
@@ -16,13 +16,14 @@ import org.openqa.selenium.Cookie
 import org.openqa.selenium.WebDriver
 import org.openqa.selenium.WebDriver.Navigation
 import org.openqa.selenium.WebElement
+import org.openqa.selenium.WindowType
 import java.net.URL
 import java.nio.file.Paths
 import java.time.LocalDateTime
 import java.time.ZoneId
-import java.time.ZoneOffset
-import java.time.ZoneOffset.*
-import java.util.*
+import java.time.ZoneOffset.UTC
+import java.util.Date
+import java.util.UUID
 import org.http4k.core.cookie.Cookie as HCookie
 
 typealias Navigate = (Request) -> Unit
@@ -46,7 +47,15 @@ class Http4kWebDriver(initialHandler: HttpHandler) : WebDriver {
     private fun navigateTo(request: Request) {
         val normalizedPath = request.uri(request.uri.path(normalized(request.uri.path)))
         val response = handler(normalizedPath)
-        current = Page(response.status, this::navigateTo, { currentUrl }, UUID.randomUUID(), normalized(latestUri), response.bodyString(), current)
+        current = Page(
+            response.status,
+            this::navigateTo,
+            { currentUrl },
+            UUID.randomUUID(),
+            normalized(latestUri),
+            response.bodyString(),
+            current
+        )
     }
 
     private fun normalized(path: String) = when {
@@ -66,7 +75,8 @@ class Http4kWebDriver(initialHandler: HttpHandler) : WebDriver {
     }
 
     private fun HCookie.toWebDriver(): Cookie = Cookie(name, value, domain, path,
-        expires?.let { Date.from(it.atZone(ZoneId.systemDefault()).toInstant()) }, secure, httpOnly)
+        expires?.let { Date.from(it.atZone(ZoneId.systemDefault()).toInstant()) }, secure, httpOnly
+    )
 
     private fun LocalCookie.toWebDriver(): StoredCookie = StoredCookie(cookie.toWebDriver(), this)
 
@@ -115,7 +125,10 @@ class Http4kWebDriver(initialHandler: HttpHandler) : WebDriver {
         override fun activeElement(): WebElement = activeElement ?: current?.firstElement()
         ?: throw NoSuchElementException("no page loaded!")
 
-        override fun window(nameOrHandle: String?): WebDriver = if (current?.handle?.toString() != nameOrHandle) throw NoSuchElementException("window with handle$nameOrHandle") else this@Http4kWebDriver
+        override fun window(nameOrHandle: String?): WebDriver =
+            if (current?.handle?.toString() != nameOrHandle) throw NoSuchElementException("window with handle$nameOrHandle") else this@Http4kWebDriver
+
+        override fun newWindow(typeHint: WindowType?) = throw FeatureNotImplementedYet
 
         override fun defaultContent(): WebDriver = this@Http4kWebDriver
     }
@@ -144,7 +157,8 @@ class Http4kWebDriver(initialHandler: HttpHandler) : WebDriver {
 
     override fun manage() = object : WebDriver.Options {
         override fun addCookie(cookie: Cookie) {
-            siteCookies[cookie.name] = StoredCookie(cookie,
+            siteCookies[cookie.name] = StoredCookie(
+                cookie,
                 LocalCookie(HCookie(cookie.name, cookie.value), LocalDateTime.now().toInstant(UTC))
             )
         }