+++ http4k-testing/webdriver/src/test/kotlin/org/http4k/webdriver/JSoupWebElementTest.kt
@@ -10,6 +10,7 @@ import org.http4k.core.Method.POST
 import org.http4k.core.Request
 import org.jsoup.Jsoup
 import org.junit.jupiter.api.Test
+import org.openqa.selenium.By
 import org.openqa.selenium.OutputType
 import org.openqa.selenium.WebElement